food/requirements.txt
Flask
Pillow
matplotlib
opencv-python